♻️ Migrate the test into the end-to-end test for getting really token to test.
